refactor: moved ApplicationForm related DTOs to ApplicationFormDtos.kt
